Mavenized project
