+++ src/main/java/me/BadBones69/CrazyAuctions/Category.java
