+++ src/main/java/me/BadBones69/CrazyAuctions/Currency/CM.java
