+++ src/main/java/me/BadBones69/CrazyAuctions/Currency/Vault.java
