+++ src/main/java/me/BadBones69/CrazyAuctions/SettingsManager.java
